fix(create-plex-channel): wait for the bot to have access
I was trying to query the guild as soon as the user hits /discord.
Since the bot was not added yet, it was missing access.

Now I'm waiting for the GUILD_CREATE gateway event that notifies me
that the plex bot was added. After this, I can safely query the channels
and then create the plex channels if it doesn't exists

Closes #5