fix(notif): fix error when creating a notification
I'm not sure how to reproduce the error since it's when Plex
doesn't send the metadata (the show was not found on thetvdb.com ?)

An easy fix it to skip the notification sending when the medatada
is null. It doesn't impact the end user since a notification without
info is useless anyway.

Closes #2